Powersense fix refactor for devices
