+++ Pulse/Models/HistoryFetcher.swift
@@ -179,11 +179,3 @@ actor ResultsMerger {
         return mergedResults
     }
 }
-
-extension Array {
-    func chunked(into size: Int) -> [[Element]] {
-        return stride(from: 0, to: count, by: size).map {
-            Array(self[$0 ..< Swift.min($0 + size, count)])
-        }
-    }
-}
