+++ Pulse/Models/SwiftData Models/PowerSenseEvent.swift
@@ -118,6 +118,89 @@ final class PowerSenseEvent {
         resolvedAt = Date()
         outageDuration = resolvedAt!.timeIntervalSince(timestamp)
     }
+    
+    // MARK: - Utilities
+
+    /// Whether this event represents an active outage (power lost and not resolved)
+    var isActiveOutage: Bool {
+        return isActive  // Active means power is off
+    }
+
+    /// Human-readable duration string
+    var durationString: String {
+        guard let duration = outageDuration else {
+            if isActiveOutage {
+                let currentDuration = Date().timeIntervalSince(timestamp)
+                return formatDuration(currentDuration) + " (ongoing)"
+            }
+            return "N/A"
+        }
+        return formatDuration(duration)
+    }
+
+    /// Format duration as human-readable string
+    private func formatDuration(_ duration: TimeInterval) -> String {
+        let hours = Int(duration) / 3600
+        let minutes = (Int(duration) % 3600) / 60
+        let seconds = Int(duration) % 60
+
+        if hours > 0 {
+            return "\(hours)h \(minutes)m \(seconds)s"
+        } else if minutes > 0 {
+            return "\(minutes)m \(seconds)s"
+        } else {
+            return "\(seconds)s"
+        }
+    }
+
+    /// Event age (time since event occurred)
+    var age: TimeInterval {
+        return Date().timeIntervalSince(timestamp)
+    }
+
+    /// Whether event is recent (within last hour)
+    var isRecent: Bool {
+        return age < 3600
+    }
+
+
+    // MARK: - Aggregation Helpers
+
+    /// Group events by time window for analysis
+    static func groupByTimeWindow(_ events: [PowerSenseEvent], windowSize: TimeInterval = 3600) -> [Date: [PowerSenseEvent]] {
+        _ = Calendar.current
+        return Dictionary(grouping: events) { event in
+            let windowStart = floor(event.timestamp.timeIntervalSince1970 / windowSize) * windowSize
+            return Date(timeIntervalSince1970: windowStart)
+        }
+    }
+
+    /// Filter events for recent outages (active events within timeframe)
+    static func recentOutages(_ events: [PowerSenseEvent], within timeframe: TimeInterval = 3600) -> [PowerSenseEvent] {
+        let cutoff = Date().addingTimeInterval(-timeframe)
+        return events.filter { event in
+            event.isActive && event.timestamp > cutoff
+        }
+    }
+
+    /// Calculate outage statistics for a collection of events
+    static func outageStatistics(for events: [PowerSenseEvent]) -> OutageStatistics {
+        let powerLostEvents = events  // All events are power-related
+        let activeOutages = powerLostEvents.filter { $0.isActive }
+        let resolvedOutages = powerLostEvents.filter { !$0.isActive && $0.outageDuration != nil }
+
+        let totalOutages = powerLostEvents.count
+        let activeCount = activeOutages.count
+        let averageDuration = resolvedOutages.compactMap { $0.outageDuration }.average
+
+        return OutageStatistics(
+            totalOutages: totalOutages,
+            activeOutages: activeCount,
+            resolvedOutages: resolvedOutages.count,
+            averageOutageDuration: averageDuration
+        )
+    }
+
 }
 
 // MARK: - PowerSense Event Properties (API Communication)
@@ -286,93 +369,6 @@ struct PowerSenseEventProperties: Decodable {
     }
 }
 
-// MARK: - Extensions
-
-extension PowerSenseEvent {
-
-    /// Whether this event represents an active outage (power lost and not resolved)
-    var isActiveOutage: Bool {
-        return isActive  // Active means power is off
-    }
-
-    /// Human-readable duration string
-    var durationString: String {
-        guard let duration = outageDuration else {
-            if isActiveOutage {
-                let currentDuration = Date().timeIntervalSince(timestamp)
-                return formatDuration(currentDuration) + " (ongoing)"
-            }
-            return "N/A"
-        }
-        return formatDuration(duration)
-    }
-
-    /// Format duration as human-readable string
-    private func formatDuration(_ duration: TimeInterval) -> String {
-        let hours = Int(duration) / 3600
-        let minutes = (Int(duration) % 3600) / 60
-        let seconds = Int(duration) % 60
-
-        if hours > 0 {
-            return "\(hours)h \(minutes)m \(seconds)s"
-        } else if minutes > 0 {
-            return "\(minutes)m \(seconds)s"
-        } else {
-            return "\(seconds)s"
-        }
-    }
-
-    /// Event age (time since event occurred)
-    var age: TimeInterval {
-        return Date().timeIntervalSince(timestamp)
-    }
-
-    /// Whether event is recent (within last hour)
-    var isRecent: Bool {
-        return age < 3600
-    }
-}
-
-// MARK: - Aggregation Helpers
-
-extension PowerSenseEvent {
-
-    /// Group events by time window for analysis
-    static func groupByTimeWindow(_ events: [PowerSenseEvent], windowSize: TimeInterval = 3600) -> [Date: [PowerSenseEvent]] {
-        _ = Calendar.current
-        return Dictionary(grouping: events) { event in
-            let windowStart = floor(event.timestamp.timeIntervalSince1970 / windowSize) * windowSize
-            return Date(timeIntervalSince1970: windowStart)
-        }
-    }
-
-    /// Filter events for recent outages (active events within timeframe)
-    static func recentOutages(_ events: [PowerSenseEvent], within timeframe: TimeInterval = 3600) -> [PowerSenseEvent] {
-        let cutoff = Date().addingTimeInterval(-timeframe)
-        return events.filter { event in
-            event.isActive && event.timestamp > cutoff
-        }
-    }
-
-    /// Calculate outage statistics for a collection of events
-    static func outageStatistics(for events: [PowerSenseEvent]) -> OutageStatistics {
-        let powerLostEvents = events  // All events are power-related
-        let activeOutages = powerLostEvents.filter { $0.isActive }
-        let resolvedOutages = powerLostEvents.filter { !$0.isActive && $0.outageDuration != nil }
-
-        let totalOutages = powerLostEvents.count
-        let activeCount = activeOutages.count
-        let averageDuration = resolvedOutages.compactMap { $0.outageDuration }.average
-
-        return OutageStatistics(
-            totalOutages: totalOutages,
-            activeOutages: activeCount,
-            resolvedOutages: resolvedOutages.count,
-            averageOutageDuration: averageDuration
-        )
-    }
-}
-
 // MARK: - OutageStatistics
 
 /// Statistics summary for PowerSense outages
@@ -385,14 +381,5 @@ struct OutageStatistics {
     var outageRate: Double {
         guard totalOutages > 0 else { return 0.0 }
         return Double(activeOutages) / Double(totalOutages)
-    }
-}
-
-// MARK: - Array Extensions
-
-private extension Array where Element == TimeInterval {
-    var average: TimeInterval? {
-        guard !isEmpty else { return nil }
-        return reduce(0, +) / Double(count)
     }
 }
