+++ Pulse/Models/SwiftData Models/SpatialClusteringActor.swift
